% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qQTLR-package.R
\docType{package}
\name{qQTLR-package}
\alias{qQTLR}
\alias{qQTLR-package}
\title{qQTLR: Quantile QTL Analysis}
\description{
Tools for quantile QTL analysis, including QUAIL vQTL mapping,
quantile TWAS weight training, and quantile TWAS pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StatFunGen/qQTLR}
  \item Report bugs at \url{https://github.com/StatFunGen/qQTLR/issues}
}

}
\author{
\strong{Maintainer}: Anjing Liu \email{anjing.liu@example.com}

Other contributors:
\itemize{
  \item Gao Wang [contributor]
}

}
\keyword{internal}
